+++ src/infra/src/cache/file_data/mod.rs
@@ -249,7 +249,7 @@ async fn download_from_storage(
         // if the downloaded length is not equal to what the blog store
         // sent in headers, we might have a partial download, so we log
         // and retry
-        if data_len != expected_blob_size {
+        if data_len as u64 != expected_blob_size {
             let msg = if i == DOWNLOAD_RETRY_TIMES - 1 {
                 format!("after {DOWNLOAD_RETRY_TIMES} retries")
             } else {
@@ -268,7 +268,7 @@ async fn download_from_storage(
     }
     // if even after retries, the download size does not match, we skip it
     // no point in validating or setting the value
-    if data_len != expected_blob_size {
+    if data_len as u64 != expected_blob_size {
         return Err(anyhow::anyhow!(
             "file {file} could not be downloaded completely: expected {expected_blob_size}, got {data_len} skipping"
         ));
@@ -344,7 +344,7 @@ pub async fn set(key: &str, data: bytes::Bytes) -> Result<(), anyhow::Error> {
 pub async fn get(
     account: &str,
     file: &str,
-    range: Option<Range<usize>>,
+    range: Option<Range<u64>>,
 ) -> object_store::Result<bytes::Bytes> {
     get_opts(account, file, range, true).await
 }
@@ -352,7 +352,7 @@ pub async fn get(
 pub async fn get_opts(
     account: &str,
     file: &str,
-    range: Option<Range<usize>>,
+    range: Option<Range<u64>>,
     remote: bool,
 ) -> object_store::Result<bytes::Bytes> {
     let cfg = config::get_config();
@@ -405,7 +405,7 @@ pub async fn get_size_opts(account: &str, file: &str, remote: bool) -> object_st
     // get from storage
     if remote {
         let meta = crate::storage::head(account, file).await?;
-        return Ok(meta.size);
+        return Ok(meta.size as usize);
     }
 
     Err(object_store::Error::NotFound {
